[REF] Conteudo do arquivo 'nfse_paulistana' movido para arquivo separado.
