+++ br_nfse/models/__init__.py
@@ -9,4 +9,5 @@ from . import res_company
 from . import nfse_susesu
 from . import nfse_simpliss
 from . import nfse_ginfes
+from . import nfse_paulistana
 # from . import account_fiscal_position
